requirements.txt
tavily-python 
fastapi
uvicorn
pydantic